pyproject.toml
[project]
name = "course-bot"
version = "0.1.0"
description = "雲科大搶課機器人"
authors = [{name = "Leo Lee", email = "leo2012c@gmail.com"}]
readme = "README.md"
requires-python = ">=3.8"

dependencies = [
    "aiofiles>=24.1.0",
    "asyncio>=4.0.0",
    "beautifulsoup4>=4.13.5",
    "dotenv>=0.9.9",
    "easyocr>=1.7.2",
    "opencv-python>=4.12.0.88",
    "pillow>=10.4.0",
    "playwright>=1.48.0",
    "pytesseract>=0.3.13",
]

[tool.uv]
dev-dependencies = [
    "pytest",
    "black",
    "flake8"
]
